fix: refactor database and utils function to remove error on render
Previous implementation of fetchData uses the wrong slice index for
pagination resulting in empty result returned on the withFetchTableData
component. Modified the operation from (page - 1) * 10 to (page * 10).

Change the database from an object with three properties of pagination
to a single array that we can use on the fetchData and findPageCount
utils function.